refactor: remove index.ts and update build process
- Deleted index.ts file, consolidating functionality into a more modular structure.
- Updated build script in package.json to use tsup for improved performance.
- Enhanced TypeScript configuration with new paths and root directory settings.
- Added new dependencies for better module resolution and build efficiency.
- Updated pnpm-lock.yaml to reflect changes in dependencies and versions.
