+++ package.json
@@ -37,9 +37,9 @@
   ],
   "scripts": {
     "dev": "tsc -p . -w",
-    "build": "rimraf ./dist && tsc -p . ",
-    "prettify": "prettier index.ts",
-    "format": "pnpm prettify --write && eslint . --fix"
+    "build": "rimraf ./dist && tsup",
+    "format": "eslint . --fix",
+    "lint": "eslint ."
   },
   "prettier": {
     "singleQuote": true,
@@ -73,6 +73,8 @@
     "eslint-plugin-prettier": "^4.2.1",
     "prettier": "^2.8.8",
     "rimraf": "^5.0.5",
+    "tsconfig-paths": "^4.2.0",
+    "tsup": "^6.7.0",
     "typescript": "^4.9.5"
   },
   "eslintConfig": {
@@ -102,7 +104,10 @@
     ],
     "settings": {
       "import/resolver": {
-        "typescript": true,
+        "typescript": {
+          "project": "./tsconfig.json",
+          "alwaysTryTypes": true
+        },
         "node": true
       }
     },
