+++ src/core/generator.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 import path from 'path';
 
 import chalk from 'chalk';
@@ -8,23 +6,15 @@ import { readFile, remove, rename, writeFile } from 'fs-extra';
 import postcss from 'postcss';
 import postcssrc from 'postcss-load-config';
 import DtsCreator from 'typed-css-modules';
-import yargs from 'yargs';
-import { hideBin } from 'yargs/helpers';
 
-type Params = {
-  verbose: boolean;
-  configPath?: string;
-  keep?: boolean;
-  namedExports: boolean;
-};
+import {
+  COMPILED_CSS_PREFIX,
+  TYPE_DEF_FILE_EXT,
+  type Params,
+} from '@/types/index';
+import { logger } from '@/utils/logger';
 
-const COMPILED_CSS_PREFIX = '__postcss__';
-const TYPE_DEF_FILE_EXT = 'd.ts';
-
-// eslint-disable-next-line @typescript-eslint/no-empty-function
-const logger = (enable: boolean) => (enable ? console.log : () => {});
-
-const run = async (
+export const run = async (
   source: string,
   { verbose, configPath, namedExports, keep = false }: Params
 ): Promise<void> => {
@@ -115,64 +105,3 @@ const run = async (
     process.exit(1);
   }
 };
-
-void yargs(hideBin(process.argv))
-  .group(['verbose', 'config', 'keep'], 'Basic options:')
-  .option('verbose', {
-    alias: 'v',
-    type: 'boolean',
-    description: 'Run with verbose logging',
-  })
-  .option('config', {
-    alias: 'c',
-    type: 'string',
-    demandOption: false,
-    description: 'Optionally provide custom path to your PostCSS config',
-  })
-  .option('keep', {
-    alias: 'k',
-    type: 'boolean',
-    demandOption: false,
-    description: 'Keep compiled css files',
-  })
-  .group(['namedExports'], 'typed-css-modules options:')
-  .option('namedExports', {
-    alias: 'n',
-    type: 'boolean',
-    default: false,
-    description: 'Enables named export for generated d.ts files',
-  })
-  .command(
-    '$0 <source> [options]',
-    'Generate d.ts files for PostCSS powered css modules',
-    (yargs) =>
-      yargs.positional('source', {
-        describe: 'source pattern to your css modules files location',
-        type: 'string',
-      }),
-    async (argv) => {
-      await run(argv.source, {
-        verbose: argv.verbose,
-        configPath: argv.config,
-        keep: argv.keep,
-        namedExports: argv.namedExports,
-      });
-    }
-  )
-  .example(
-    'pcssmdts "src/**/*.module.css"',
-    'Basic usage, in the end next to each found css file corresponding d.ts file will be generated\n\n'
-  )
-  .example(
-    'pcssmdts "src/**/*.module.css" -k',
-    'In this case compiled files will be preserved, all compiled files are prefixed by __postcss__\n\n'
-  )
-  .example(
-    'pcssmdts "src/**/*.module.css" -c configs/postcss.config.js',
-    'Custom PostCSS config location path\n\n'
-  )
-  .example(
-    'pcssmdts "src/**/*.module.css" -n',
-    'Named exports is used for generated files\n\n'
-  )
-  .parse();
