+++ src/index.ts
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+export * from '@/cli';
