+++ src/types/index.d.ts
@@ -1,0 +1,9 @@
+export type Params = {
+  verbose: boolean;
+  configPath?: string;
+  keep?: boolean;
+  namedExports: boolean;
+};
+
+export const COMPILED_CSS_PREFIX = '__postcss__';
+export const TYPE_DEF_FILE_EXT = 'd.ts';
