+++ tsconfig.json
@@ -3,14 +3,21 @@
     "baseUrl": ".",
     "outDir": "dist",
     "module": "CommonJS",
-    "target": "ES5",
+    "target": "ES2021",
     "moduleResolution": "Node",
     "esModuleInterop": true,
     "skipLibCheck": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "typeRoots": ["./node_modules/@types"]
+    "typeRoots": ["./node_modules/@types"],
+    "paths": {
+      "@/*": ["./src/*"]
+    },
+    "rootDir": "."
   },
   "exclude": ["node_modules", "dist"],
-  "include": ["index.ts"]
+  "include": ["src", "tsup.config.ts"],
+  "ts-node": {
+    "require": ["tsconfig-paths/register"]
+  }
 }
